refactor: extract local artifact storage
