+++ crates/kache-store/Cargo.toml
@@ -1,0 +1,47 @@
+[package]
+name = "kache-store"
+version = "0.16.1"
+edition = "2024"
+rust-version = "1.95"
+description = "Local content-addressed artifact storage for kache."
+license = "Apache-2.0"
+repository = "https://github.com/kunobi-ninja/kache"
+
+[features]
+test-support = []
+
+[dependencies]
+kache-format = { version = "0.16.1", path = "../kache-format" }
+anyhow = "1"
+rusqlite = { version = "0.40", features = ["bundled"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+blake3 = "1"
+libc = "0.2"
+tracing = "0.1"
+filetime = "0.2"
+tempfile = "3"
+
+[dev-dependencies]
+tempfile = "3"
+serde_json = "1"
+
+[target.'cfg(windows)'.dependencies]
+windows-sys = { version = "0.61", features = [
+    "Win32_Foundation",
+    "Win32_System_Threading",
+    "Win32_Storage_FileSystem",
+    # ReFS block-cloning (FSCTL_DUPLICATE_EXTENTS_TO_FILE) for zero-copy,
+    # write-isolated restore on a Dev Drive — see link::try_reflink (#429).
+    "Win32_System_IO",
+    "Win32_System_Ioctl",
+    # FILE_SUPPORTS_BLOCK_REFCOUNTING — ask the volume whether it block-clones
+    # instead of inferring "no CoW" from one failed clone (#508).
+    "Win32_System_SystemServices",
+] }
+blake3 = { version = "1", features = ["pure"] }
+
+# Preserve the lint settings of the binary these modules came from.
+[lints.clippy]
+permissions_set_readonly_false = "allow"
+too_many_arguments = "allow"
