+++ crates/kache-store/src/sharing.rs
