+++ src/cache_key.rs
@@ -2,7 +2,7 @@ use crate::args::RustcArgs;
 use crate::path_normalizer::{PathNormalizer, check_for_path_leak};
 use anyhow::{Context, Result};
 pub(crate) use kache_format::{is_valid_cache_key, is_valid_crate_name};
-use rusqlite::{Connection, OptionalExtension, params};
+pub(crate) use kache_store::file_hash::*;
 use std::borrow::Cow;
 use std::cell::{Cell, RefCell};
 use std::collections::{BTreeMap, HashMap};
@@ -247,7 +247,6 @@ use std::path::{Path, PathBuf};
 // unprobed. Existing Windows linked-output keys did not contain this
 // identity, so invalidate them rather than mix schemas.
 pub(crate) use kache_format::CACHE_KEY_VERSION;
-const MIN_PERSISTED_HASH_BYTES: i64 = 64 * 1024;
 
 /// Collapse runs of ASCII whitespace into single spaces and trim
 /// leading / trailing whitespace.
@@ -2671,17 +2670,6 @@ fn is_ident_continue(byte: u8) -> bool {
 // substitution). The ad-hoc helper had two failure modes — see
 // the `path_normalizer` module docs for the full story.
 
-/// Hash a file using blake3.
-pub fn hash_file(path: &Path) -> Result<String> {
-    let file = std::fs::File::open(path)
-        .with_context(|| format!("opening {} for hashing", path.display()))?;
-    let mut hasher = blake3::Hasher::new();
-    hasher
-        .update_reader(file)
-        .with_context(|| format!("reading {} for hashing", path.display()))?;
-    Ok(hasher.finalize().to_hex().to_string())
-}
-
 /// Compute a linked `static=` archive's cache-key digest. A proven GNU/BSD
 /// archive gets the structural member-identity hash; every other non-thin
 /// archive gets a digest of both its bytes and lexical absolute path because
@@ -3089,12 +3077,6 @@ struct FileHashStatsCells {
     bytes_hashed: Cell<u64>,
 }
 
-enum FileHashCache<'db> {
-    Borrowed(&'db Connection),
-    #[cfg(test)]
-    Owned(Connection),
-}
-
 /// One input of a memoised preprocessor run: its metadata at the time the
 /// expansion was hashed, plus the hash of its contents.
 ///
@@ -3135,31 +3117,6 @@ impl CcPreprocessMemoInput {
     pub(crate) fn local_path(&self) -> &str {
         &self.fingerprint.path
     }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize)]
-pub(crate) struct FileFingerprint {
-    path: String,
-    size: i64,
-    mtime_ns: i64,
-    ctime_ns: i64,
-    /// Filesystem inode (0 on non-Unix / unavailable). Folded into the memo key
-    /// so an in-place swap that preserves path+size+mtime+ctime but changes the
-    /// inode (and content) can't return a stale memoized hash (kunobi-ninja/kache#324).
-    inode: i64,
-}
-
-/// Result of a content-hash cache lookup that does NOT compute a blake3 — the
-/// lock-narrowing seam for the daemon's `HashFiles` path (#281). The caller
-/// hashes (`hash_file`) outside any store lock on a miss, then records via
-/// [`FileHasher::record_cached`].
-pub(crate) enum FileHashLookup {
-    /// Cached hash found — no hashing needed.
-    Hit(String),
-    /// Cache miss; hash the file then record under this fingerprint.
-    NeedsHash(FileFingerprint),
-    /// Too small to persist, or metadata unreadable — hash but don't cache.
-    Uncacheable,
 }
 
 #[derive(Debug, Clone)]
@@ -3216,9 +3173,9 @@ impl FileHasher<'static> {
 }
 
 impl<'db> FileHasher<'db> {
-    pub(crate) fn from_connection(db: &'db Connection) -> Self {
+    pub(crate) fn from_cache(cache: FileHashCache<'db>) -> Self {
         FileHasher {
-            cache: Some(FileHashCache::Borrowed(db)),
+            cache: Some(cache),
             daemon_socket: None,
             use_input_predictions: false,
             prefetched: RefCell::new(HashMap::new()),
@@ -3807,70 +3764,6 @@ impl<'db> FileHasher<'db> {
         Ok(result)
     }
 
-    /// Cache lookup ONLY — reads the persistent hash cache, never computes a
-    /// blake3. Lets a caller holding a coarse lock (the daemon's `Mutex<Store>`)
-    /// release it before the expensive file read and re-take it only for the
-    /// short record (#281). Mirrors [`Self::hash`]'s fingerprint + min-size +
-    /// cache-get logic exactly, so the cache key is identical.
-    pub(crate) fn lookup_cached(&self, path: &Path) -> FileHashLookup {
-        let Some(cache) = &self.cache else {
-            return FileHashLookup::Uncacheable;
-        };
-        let fingerprint = match FileFingerprint::from_path(path) {
-            Ok(fp) => fp,
-            Err(e) => {
-                tracing::debug!(
-                    "file hash cache metadata lookup failed for {}: {e}",
-                    path.display()
-                );
-                return FileHashLookup::Uncacheable;
-            }
-        };
-        if fingerprint.size < MIN_PERSISTED_HASH_BYTES {
-            return FileHashLookup::Uncacheable;
-        }
-        match cache.get(&fingerprint) {
-            Ok(Some(hash)) => FileHashLookup::Hit(hash),
-            Ok(None) => FileHashLookup::NeedsHash(fingerprint),
-            Err(e) => {
-                // Treat a lookup error as a miss — recompute rather than fail.
-                tracing::debug!("file hash cache lookup failed for {}: {e}", path.display());
-                FileHashLookup::NeedsHash(fingerprint)
-            }
-        }
-    }
-
-    /// Record a freshly-computed hash for `fingerprint` (the miss arm of
-    /// [`Self::lookup_cached`]). Best-effort — a cache write failure is logged,
-    /// not propagated.
-    pub(crate) fn record_cached(&self, fingerprint: &FileFingerprint, hash: &str) {
-        if let Some(cache) = &self.cache
-            && let Err(e) = cache.put(fingerprint, hash)
-        {
-            tracing::debug!("file hash cache update failed: {e}");
-        }
-    }
-
-    /// Record a hash the caller already knows for `fingerprint` — without
-    /// reading the file (kunobi-ninja/kache#540).
-    ///
-    /// The caller must have established that hash for THAT fingerprint, not
-    /// merely for that path: the row is only ever served back on an exact
-    /// fingerprint match, so a fingerprint captured at the moment the content
-    /// was known stays a true statement even if the file changes a moment
-    /// later — the changed file simply misses and gets hashed. Re-stating the
-    /// path here instead would pair the new file's fingerprint with the old
-    /// file's hash.
-    ///
-    /// Honors the same size floor as [`Self::hash`], which would not consult
-    /// the memo for a smaller file anyway.
-    pub(crate) fn record_verified(&self, fingerprint: &FileFingerprint, hash: &str) {
-        if fingerprint.size < MIN_PERSISTED_HASH_BYTES {
-            return;
-        }
-        self.record_cached(fingerprint, hash);
-    }
-
     /// Hash a linked `-l static=` archive for the cache key. Clean GNU/BSD
     /// archives use a structural digest that retains exact member identity.
     /// Other non-thin inputs use a path-bound fallback; thin archives error so
@@ -3962,305 +3855,6 @@ impl<'db> FileHasher<'db> {
             .bytes_hashed
             .set(self.stats.bytes_hashed.get().saturating_add(bytes));
     }
-}
-
-impl<'db> FileHashCache<'db> {
-    #[cfg(test)]
-    fn open(index_db_path: &Path) -> Result<Self> {
-        let db = Connection::open(index_db_path)
-            .with_context(|| format!("opening file hash cache {}", index_db_path.display()))?;
-        db.pragma_update(None, "busy_timeout", "5000")?;
-        db.pragma_update(None, "journal_mode", "WAL")?;
-        db.pragma_update(None, "synchronous", "NORMAL")?;
-        ensure_file_hash_cache_schema(&db)?;
-        Ok(Self::Owned(db))
-    }
-
-    fn db(&self) -> &Connection {
-        match self {
-            Self::Borrowed(db) => db,
-            #[cfg(test)]
-            Self::Owned(db) => db,
-        }
-    }
-
-    fn get(&self, fingerprint: &FileFingerprint) -> rusqlite::Result<Option<String>> {
-        self.db()
-            .query_row(
-                "SELECT hash FROM file_hashes
-                 WHERE path = ?1 AND size = ?2 AND mtime_ns = ?3 AND ctime_ns = ?4 AND inode = ?5",
-                params![
-                    fingerprint.path,
-                    fingerprint.size,
-                    fingerprint.mtime_ns,
-                    fingerprint.ctime_ns,
-                    fingerprint.inode
-                ],
-                |row| row.get(0),
-            )
-            .optional()
-    }
-
-    fn put(&self, fingerprint: &FileFingerprint, hash: &str) -> rusqlite::Result<()> {
-        self.db().execute(
-            "INSERT OR REPLACE INTO file_hashes
-             (path, size, mtime_ns, ctime_ns, inode, hash, updated_at)
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6, datetime('now'))",
-            params![
-                fingerprint.path,
-                fingerprint.size,
-                fingerprint.mtime_ns,
-                fingerprint.ctime_ns,
-                fingerprint.inode,
-                hash
-            ],
-        )?;
-        Ok(())
-    }
-
-    fn get_runtime_env_use(&self, content_hash: &str, var: &str) -> rusqlite::Result<Option<bool>> {
-        self.db()
-            .query_row(
-                "SELECT has_runtime_use FROM source_env_runtime_uses
-                 WHERE content_hash = ?1 AND env_var = ?2",
-                params![content_hash, var],
-                |row| row.get::<_, i64>(0).map(|value| value != 0),
-            )
-            .optional()
-    }
-
-    fn put_runtime_env_use(
-        &self,
-        content_hash: &str,
-        var: &str,
-        has_runtime_use: bool,
-    ) -> rusqlite::Result<()> {
-        self.db().execute(
-            "INSERT OR REPLACE INTO source_env_runtime_uses
-             (content_hash, env_var, has_runtime_use, updated_at)
-             VALUES (?1, ?2, ?3, datetime('now'))",
-            params![content_hash, var, i64::from(has_runtime_use)],
-        )?;
-        Ok(())
-    }
-
-    fn get_cc_preprocess_memo(&self, memo_key: &str) -> rusqlite::Result<Option<(String, String)>> {
-        self.db()
-            .query_row(
-                "SELECT preprocessed_hash, inputs_json FROM cc_preprocess_memos
-                 WHERE memo_key = ?1",
-                params![memo_key],
-                |row| Ok((row.get(0)?, row.get(1)?)),
-            )
-            .optional()
-    }
-
-    /// The recorded closure for `identity`, or `None` when there is no row.
-    ///
-    /// A row written by a schema this build does not recognise reads as
-    /// absent: the cost of that is one pre-pass, and the cost of guessing at
-    /// an unknown encoding is a wrong key.
-    fn get_input_prediction(&self, identity: &str) -> rusqlite::Result<Option<(u32, String)>> {
-        self.db()
-            .query_row(
-                "SELECT schema, prediction_json FROM input_predictions WHERE identity = ?1",
-                params![identity],
-                |row| Ok((row.get(0)?, row.get(1)?)),
-            )
-            .optional()
-    }
-
-    fn put_input_prediction(
-        &self,
-        identity: &str,
-        schema: u32,
-        crate_name: Option<&str>,
-        prediction_json: &str,
-    ) -> rusqlite::Result<()> {
-        self.db().execute(
-            "INSERT OR REPLACE INTO input_predictions
-             (identity, schema, crate_name, prediction_json, updated_at)
-             VALUES (?1, ?2, ?3, ?4, datetime('now'))",
-            params![identity, schema, crate_name, prediction_json],
-        )?;
-        Ok(())
-    }
-
-    fn put_cc_preprocess_memo(
-        &self,
-        memo_key: &str,
-        preprocessed_hash: &str,
-        inputs_json: &str,
-    ) -> rusqlite::Result<()> {
-        self.db().execute(
-            "INSERT OR REPLACE INTO cc_preprocess_memos
-             (memo_key, preprocessed_hash, inputs_json, updated_at)
-             VALUES (?1, ?2, ?3, datetime('now'))",
-            params![memo_key, preprocessed_hash, inputs_json],
-        )?;
-        Ok(())
-    }
-}
-
-pub(crate) fn ensure_file_hash_cache_schema(db: &Connection) -> rusqlite::Result<()> {
-    db.execute_batch(
-        "CREATE TABLE IF NOT EXISTS file_hashes (
-            path       TEXT PRIMARY KEY,
-            size       INTEGER NOT NULL,
-            mtime_ns   INTEGER NOT NULL,
-            ctime_ns   INTEGER NOT NULL DEFAULT 0,
-            inode      INTEGER NOT NULL DEFAULT 0,
-            hash       TEXT NOT NULL,
-            updated_at TEXT NOT NULL DEFAULT (datetime('now'))
-        );
-        CREATE TABLE IF NOT EXISTS cc_preprocess_memos (
-            memo_key          TEXT PRIMARY KEY,
-            preprocessed_hash TEXT NOT NULL,
-            inputs_json       TEXT NOT NULL,
-            updated_at        TEXT NOT NULL DEFAULT (datetime('now'))
-        );
-        CREATE TABLE IF NOT EXISTS input_predictions (
-            identity        TEXT PRIMARY KEY,
-            schema          INTEGER NOT NULL,
-            crate_name      TEXT,
-            prediction_json TEXT NOT NULL,
-            updated_at      TEXT NOT NULL DEFAULT (datetime('now'))
-        );
-        CREATE TABLE IF NOT EXISTS source_env_runtime_uses (
-            content_hash    TEXT NOT NULL,
-            env_var         TEXT NOT NULL,
-            has_runtime_use INTEGER NOT NULL,
-            updated_at      TEXT NOT NULL DEFAULT (datetime('now')),
-            PRIMARY KEY (content_hash, env_var)
-        );",
-    )?;
-    for column in [
-        "ALTER TABLE file_hashes ADD COLUMN ctime_ns INTEGER NOT NULL DEFAULT 0",
-        "ALTER TABLE file_hashes ADD COLUMN inode INTEGER NOT NULL DEFAULT 0",
-    ] {
-        if let Err(e) = db.execute_batch(column)
-            && !e.to_string().contains("duplicate column name")
-        {
-            return Err(e);
-        }
-    }
-    Ok(())
-}
-
-impl FileFingerprint {
-    /// Identity of a file as the memo sees it. Also the cheapest available
-    /// proof that an external tool did NOT rewrite a file across some
-    /// operation: any in-place write bumps `mtime_ns`/`ctime_ns` (and a
-    /// replace-by-rename changes `inode`), so an unchanged fingerprint means
-    /// unchanged bytes. `restore_from_cache` reads it that way (#540).
-    pub(crate) fn from_path(path: &Path) -> Result<Self> {
-        let metadata = std::fs::metadata(path)
-            .with_context(|| format!("reading metadata for {}", path.display()))?;
-        let absolute_path = absolute_path(path);
-
-        Ok(Self {
-            path: absolute_path.to_string_lossy().into_owned(),
-            size: i64::try_from(metadata.len()).unwrap_or(i64::MAX),
-            mtime_ns: metadata_mtime_ns(&metadata),
-            ctime_ns: metadata_ctime_ns(&metadata),
-            inode: metadata_inode(&metadata),
-        })
-    }
-}
-
-fn absolute_path(path: &Path) -> PathBuf {
-    if path.is_absolute() {
-        path.to_path_buf()
-    } else {
-        std::env::current_dir()
-            .map(|cwd| cwd.join(path))
-            .unwrap_or_else(|_| path.to_path_buf())
-    }
-}
-
-/// Filesystem inode number (0 where unavailable, e.g. non-Unix).
-pub(crate) fn metadata_inode(metadata: &std::fs::Metadata) -> i64 {
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::MetadataExt;
-        i64::try_from(metadata.ino()).unwrap_or(i64::MAX)
-    }
-    #[cfg(not(unix))]
-    {
-        let _ = metadata;
-        0
-    }
-}
-
-pub(crate) fn metadata_mtime_ns(metadata: &std::fs::Metadata) -> i64 {
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::MetadataExt;
-
-        metadata_parts_ns(metadata.mtime(), metadata.mtime_nsec())
-    }
-
-    #[cfg(windows)]
-    {
-        use std::os::windows::fs::MetadataExt;
-
-        windows_filetime_ns(metadata.last_write_time())
-    }
-
-    #[cfg(not(any(unix, windows)))]
-    {
-        system_time_ns(metadata.modified().ok()).unwrap_or_default()
-    }
-}
-
-pub(crate) fn metadata_ctime_ns(metadata: &std::fs::Metadata) -> i64 {
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::MetadataExt;
-
-        metadata_parts_ns(metadata.ctime(), metadata.ctime_nsec())
-    }
-
-    #[cfg(windows)]
-    {
-        use std::os::windows::fs::MetadataExt;
-
-        windows_filetime_ns(metadata.creation_time())
-    }
-
-    #[cfg(not(any(unix, windows)))]
-    {
-        system_time_ns(metadata.created().ok()).unwrap_or_else(|| metadata_mtime_ns(metadata))
-    }
-}
-
-#[cfg(unix)]
-fn metadata_parts_ns(seconds: i64, nanoseconds: i64) -> i64 {
-    seconds
-        .saturating_mul(1_000_000_000)
-        .saturating_add(nanoseconds)
-}
-
-#[cfg(windows)]
-fn windows_filetime_ns(filetime_100ns: u64) -> i64 {
-    const UNIX_EPOCH_FILETIME_100NS: u64 = 116_444_736_000_000_000;
-
-    filetime_100ns
-        .saturating_sub(UNIX_EPOCH_FILETIME_100NS)
-        .saturating_mul(100)
-        .min(i64::MAX as u64) as i64
-}
-
-#[cfg(not(any(unix, windows)))]
-fn system_time_ns(time: Option<std::time::SystemTime>) -> Option<i64> {
-    let duration = time?.duration_since(std::time::UNIX_EPOCH).ok()?;
-    let seconds = i64::try_from(duration.as_secs()).unwrap_or(i64::MAX / 1_000_000_000);
-
-    Some(
-        seconds
-            .saturating_mul(1_000_000_000)
-            .saturating_add(i64::from(duration.subsec_nanos())),
-    )
 }
 
 /// `-C extra-filename=` in either spelling. rustc normalises `-` and `_` in
@@ -7516,40 +7110,6 @@ mod tests {
     }
 
     #[test]
-    fn test_hash_file() {
-        let dir = tempfile::tempdir().unwrap();
-        let file = dir.path().join("test.rs");
-        std::fs::write(&file, b"fn main() {}").unwrap();
-
-        let hash = hash_file(&file).unwrap();
-        assert_eq!(hash.len(), 64); // blake3 hex is 64 chars
-
-        // Same content = same hash
-        let file2 = dir.path().join("test2.rs");
-        std::fs::write(&file2, b"fn main() {}").unwrap();
-        let hash2 = hash_file(&file2).unwrap();
-        assert_eq!(hash, hash2);
-
-        // Different content = different hash
-        let file3 = dir.path().join("test3.rs");
-        std::fs::write(&file3, b"fn main() { println!(\"hello\"); }").unwrap();
-        let hash3 = hash_file(&file3).unwrap();
-        assert_ne!(hash, hash3);
-
-        // Larger than blake3's streaming read buffer so the digest spans
-        // multiple reads instead of relying on a single in-memory buffer.
-        let large = dir.path().join("large.rlib");
-        let large_bytes: Vec<u8> = (0..256 * 1024 + 17)
-            .map(|index| (index % 251) as u8)
-            .collect();
-        std::fs::write(&large, &large_bytes).unwrap();
-        assert_eq!(
-            hash_file(&large).unwrap(),
-            blake3::hash(&large_bytes).to_hex().to_string()
-        );
-    }
-
-    #[test]
     fn test_cache_key_deterministic() {
         let _lock = key_test_lock();
         let dir = tempfile::tempdir().unwrap();
@@ -9769,35 +9329,6 @@ pub const OUT_DIR_AT_COMPILE_TIME: &str = env!("OUT_DIR");
     }
 
     #[test]
-    fn file_hash_memo_key_includes_inode() {
-        // An in-place swap that preserves path+size+mtime+ctime but changes the
-        // inode (and content) must NOT return a stale memoized hash
-        // (kunobi-ninja/kache#324).
-        let conn = rusqlite::Connection::open_in_memory().unwrap();
-        ensure_file_hash_cache_schema(&conn).unwrap();
-        let cache = FileHashCache::Borrowed(&conn);
-
-        let fp = |inode: i64| FileFingerprint {
-            path: "/x/lib.rlib".to_string(),
-            size: 100,
-            mtime_ns: 1,
-            ctime_ns: 2,
-            inode,
-        };
-
-        cache.put(&fp(10), "hash_for_inode_10").unwrap();
-        assert_eq!(
-            cache.get(&fp(10)).unwrap().as_deref(),
-            Some("hash_for_inode_10")
-        );
-        assert_eq!(
-            cache.get(&fp(20)).unwrap(),
-            None,
-            "a different inode (same path/size/mtime/ctime) must miss the memo"
-        );
-    }
-
-    #[test]
     fn cc_preprocess_memo_requires_every_input_fingerprint_to_match() {
         let dir = tempfile::tempdir().unwrap();
         let db = dir.path().join("idx.sqlite");
@@ -10279,102 +9810,6 @@ pub const OUT_DIR_AT_COMPILE_TIME: &str = env!("OUT_DIR");
         std::fs::write(&file, vec![2u8; 70 * 1024]).unwrap();
         let changed = FileHasher::persistent(&db_path).hash(&file).unwrap();
         assert_ne!(first, changed);
-    }
-
-    #[test]
-    fn lookup_cached_without_a_cache_is_uncacheable() {
-        // FileHasher::new() has no persistent cache, so the lock-narrowing
-        // lookup path (#281) short-circuits to Uncacheable and record_cached
-        // is a no-op. Covers the `cache: None` arms of both methods.
-        let dir = tempfile::tempdir().unwrap();
-        let file = dir.path().join("large.rlib");
-        std::fs::write(&file, vec![3u8; 70 * 1024]).unwrap();
-
-        let fh = FileHasher::new();
-        assert!(matches!(
-            fh.lookup_cached(&file),
-            FileHashLookup::Uncacheable
-        ));
-        // record_cached on a cacheless hasher must not panic and is a no-op.
-        if let FileHashLookup::NeedsHash(fp) = fh.lookup_cached(&file) {
-            fh.record_cached(&fp, "deadbeef");
-        }
-    }
-
-    #[test]
-    fn lookup_cached_too_small_or_unreadable_is_uncacheable() {
-        // A sub-threshold file and an unreadable path both yield Uncacheable
-        // from a *persistent* hasher: the first via the min-size guard, the
-        // second via the metadata-read failure arm.
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("index.db");
-        let small = dir.path().join("small.rs");
-        std::fs::write(&small, b"fn main() {}").unwrap();
-
-        let fh = FileHasher::persistent(&db_path);
-        assert!(matches!(
-            fh.lookup_cached(&small),
-            FileHashLookup::Uncacheable
-        ));
-        // Nonexistent path -> FileFingerprint::from_path errors -> Uncacheable.
-        assert!(matches!(
-            fh.lookup_cached(&dir.path().join("nope.rlib")),
-            FileHashLookup::Uncacheable
-        ));
-    }
-
-    #[test]
-    fn lookup_cached_miss_then_record_then_hit_roundtrips() {
-        // The daemon's lock-narrowing seam: a large file first reports
-        // NeedsHash (miss), then after record_cached() a subsequent
-        // lookup_cached() returns Hit with the recorded digest — without ever
-        // computing a blake3 in lookup_cached itself. Covers NeedsHash, the
-        // record_cached put arm, and the Hit arm.
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("index.db");
-        let file = dir.path().join("large.rlib");
-        std::fs::write(&file, vec![7u8; 70 * 1024]).unwrap();
-
-        let fh = FileHasher::persistent(&db_path);
-        let fp = match fh.lookup_cached(&file) {
-            FileHashLookup::NeedsHash(fp) => fp,
-            _ => panic!("expected NeedsHash on first lookup"),
-        };
-        fh.record_cached(&fp, "cafef00d");
-
-        match fh.lookup_cached(&file) {
-            FileHashLookup::Hit(h) => assert_eq!(h, "cafef00d"),
-            _ => panic!("expected Hit after record_cached"),
-        }
-    }
-
-    #[test]
-    fn record_verified_honors_the_persistence_floor() {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("index.db");
-        let below = dir.path().join("below.rlib");
-        let at_floor = dir.path().join("at-floor.rlib");
-        std::fs::write(&below, vec![1u8; MIN_PERSISTED_HASH_BYTES as usize - 1]).unwrap();
-        std::fs::write(&at_floor, vec![2u8; MIN_PERSISTED_HASH_BYTES as usize]).unwrap();
-
-        let hasher = FileHasher::persistent(&db_path);
-        hasher.record_verified(&FileFingerprint::from_path(&below).unwrap(), "below");
-        hasher.record_verified(&FileFingerprint::from_path(&at_floor).unwrap(), "at-floor");
-
-        assert!(matches!(
-            hasher.lookup_cached(&below),
-            FileHashLookup::Uncacheable
-        ));
-        assert!(matches!(
-            hasher.lookup_cached(&at_floor),
-            FileHashLookup::Hit(hash) if hash == "at-floor"
-        ));
-
-        let conn = rusqlite::Connection::open(&db_path).unwrap();
-        let rows: i64 = conn
-            .query_row("SELECT COUNT(*) FROM file_hashes", [], |row| row.get(0))
-            .unwrap();
-        assert_eq!(rows, 1, "sub-threshold fingerprints must not be stored");
     }
 
     #[test]
